only make usagelog have a bigint id, and on sqlite it is integer because sqlite cant handle autoincrement on bigint. Keep other logs at integer id type as usagelog is the only one that is even close to needing this

+++ fits_storage/server/orm/downloadlog.py
@@ -13,7 +13,7 @@ class DownloadLog(Base):
     """
     __tablename__ = 'downloadlog'
 
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     usagelog_id = Column(BigInteger, ForeignKey(UsageLog.id), nullable=False,
                          index=True)
     usagelog = relationship(UsageLog, order_by=id)
